Add successive pose estimation using ICP

+++ src/pose_graph.rs
@@ -4,11 +4,11 @@ use ndarray::prelude::*;
 pub struct Edge {
     source_id: usize,
     target_id: usize,
-    information: Array2<f32>,
+    information: Array2<f64>,
 }
 
 pub struct PoseGraph {
-    pub nodes: Vec<Array2<f32>>, // 3x3 homogeneous pose matrix for 2D slam
+    pub nodes: Vec<Array2<f64>>, // 3x3 homogeneous pose matrix for 2D slam
     pub edges: Vec<Edge>,
 }
 
@@ -19,8 +19,8 @@ impl PoseGraph {
         }
 
         let pose = array![
-            [pose.theta.cos(), -pose.theta.sin(), pose.x],
-            [pose.theta.sin(), pose.theta.cos(), pose.y],
+            [pose.theta.cos() as f64, -pose.theta.sin() as f64, pose.x as f64],
+            [pose.theta.sin() as f64, pose.theta.cos() as f64, pose.y as f64],
             [0., 0., 1.],
         ];
         self.nodes.push(pose);
